[#89] refactor: formatDate 함수를 date.ts로 통합
